refactor: streamline segment processing and metadata management
- Consolidated segment text conversion to chunks into a dedicated method, improving code organization and readability.
- Enhanced the AddSegments and UpdateSegments methods to utilize the new chunk conversion method, ensuring consistent processing.
- Implemented metadata merging logic during segment updates to preserve existing metadata while incorporating new data.
- Improved error handling and logging throughout segment operations, providing better insights into processing outcomes.
