+++ graphrag/segment.go
@@ -3,763 +3,14 @@ package graphrag
 import (
 	"context"
 	"fmt"
-	"time"
 
 	"github.com/yaoapp/gou/graphrag/types"
 	"github.com/yaoapp/gou/graphrag/utils"
 )
 
 // ================================================================================================
-// CRUD Operations - Exported Methods
+// Exported Utility Functions - Segment Text to Chunk Conversion
 // ================================================================================================
-
-// AddSegments adds segments to a collection manually
-func (g *GraphRag) AddSegments(ctx context.Context, docID string, segmentTexts []types.SegmentText, options *types.UpsertOptions) ([]string, error) {
-	// Step 1: Parse CollectionID from docID
-	collectionID, _ := utils.ExtractCollectionIDFromDocID(docID)
-	if collectionID == "" {
-		collectionID = "default"
-	}
-
-	// Step 2: Prepare options by copying and setting necessary fields
-	opts := &types.UpsertOptions{}
-	if options != nil {
-		*opts = *options
-	}
-	opts.CollectionID = collectionID
-	opts.DocID = docID
-
-	// Step 3: Get collection IDs for vector and graph storage
-	collectionIDs, err := utils.GetCollectionIDs(collectionID)
-	if err != nil {
-		return nil, fmt.Errorf("failed to generate collection IDs: %w", err)
-	}
-
-	// Step 4: Convert SegmentTexts to Chunks
-	chunks, err := g.convertSegmentTextsToChunks(segmentTexts, docID)
-	if err != nil {
-		return nil, fmt.Errorf("failed to convert segment texts to chunks: %w", err)
-	}
-
-	// Step 5: Create the callback for progress tracking
-	var embeddingTexts []string = []string{}
-	var embeddingIndexesMap map[*types.Chunk]int = map[*types.Chunk]int{}
-	var cb = MakeUpsertCallback(docID, nil, opts.Progress)
-
-	// Step 6: Require embedding and extraction configurations
-	if opts.Embedding == nil {
-		return nil, fmt.Errorf("embedding configuration is required for AddSegments operation")
-	}
-
-	if g.Graph != nil && opts.Extraction == nil {
-		return nil, fmt.Errorf("extraction configuration is required when graph store is configured")
-	}
-
-	// Step 7: Store embedding indexes for chunks (equivalent to AddFile step 4.2)
-	for _, chunk := range chunks {
-		embeddingTexts = append(embeddingTexts, chunk.Text)
-		embeddingIndexesMap[chunk] = len(embeddingTexts) - 1
-	}
-
-	// Step 8: Extract entities and relationships from chunks (equivalent to AddFile step 4.3)
-	allEntities, allRelationships, entityIndexMap, relationshipIndexMap, err := g.extractEntitiesAndRelationships(ctx, chunks, opts, cb, &embeddingTexts)
-	if err != nil {
-		return nil, fmt.Errorf("failed to extract entities and relationships: %w", err)
-	}
-
-	// Step 9: Embed all texts (chunks + entities + relationships) (equivalent to AddFile step 5)
-	embeddings, err := opts.Embedding.EmbedDocuments(ctx, embeddingTexts, cb.Embedding)
-	if err != nil {
-		return nil, fmt.Errorf("failed to embed the documents: %w", err)
-	}
-
-	// Step 10: Store entities and relationships to graph store first for deduplication (equivalent to AddFile step 6)
-	var actualEntityIDs []string
-	var actualRelationshipIDs []string
-	var entityIDMap = make(map[string]string)
-	var relationshipIDMap = make(map[string]string)
-
-	var entityDeduplicationResults map[string]*EntityDeduplicationResult
-	var relationshipDeduplicationResults map[string]*RelationshipDeduplicationResult
-
-	if g.Graph != nil && opts.Extraction != nil && (len(allEntities) > 0 || len(allRelationships) > 0) {
-		// Store entities to graph store
-		if len(allEntities) > 0 {
-			actualEntityIDs, entityDeduplicationResults, err = g.storeEntitiesToGraphStore(ctx, allEntities, collectionIDs.Graph, docID)
-			if err != nil {
-				return nil, fmt.Errorf("failed to store entities to graph store: %w", err)
-			}
-
-			// Create mapping from original IDs to actual IDs
-			for i, entity := range allEntities {
-				if i < len(actualEntityIDs) {
-					entityIDMap[entity.ID] = actualEntityIDs[i]
-				}
-			}
-		}
-
-		// Store relationships to graph store
-		if len(allRelationships) > 0 {
-			actualRelationshipIDs, relationshipDeduplicationResults, err = g.storeRelationshipsToGraphStore(ctx, allRelationships, collectionIDs.Graph, docID)
-			if err != nil {
-				return nil, fmt.Errorf("failed to store relationships to graph store: %w", err)
-			}
-
-			// Create mapping from original IDs to actual IDs
-			for i, relationship := range allRelationships {
-				if i < len(actualRelationshipIDs) {
-					relationshipIDMap[relationship.ID] = actualRelationshipIDs[i]
-				}
-			}
-		}
-
-		// Update chunks with actual IDs from graph database
-		g.updateChunksWithActualIds(chunks, entityIDMap, relationshipIDMap)
-	}
-
-	// Step 11: Store all documents to vector store (equivalent to AddFile step 7)
-	storeOptions := &StoreDocumentsOptions{
-		Chunks:                           chunks,
-		Entities:                         allEntities,
-		Relationships:                    allRelationships,
-		Embeddings:                       embeddings,
-		EmbeddingIndexesMap:              embeddingIndexesMap,
-		EntityIndexMap:                   entityIndexMap,
-		RelationshipIndexMap:             relationshipIndexMap,
-		SourceFile:                       "",
-		ConvertMetadata:                  make(map[string]interface{}),
-		UserMetadata:                     opts.Metadata,
-		VectorCollectionName:             collectionIDs.Vector,
-		CollectionID:                     collectionID,
-		DocID:                            docID,
-		EntityDeduplicationResults:       entityDeduplicationResults,
-		RelationshipDeduplicationResults: relationshipDeduplicationResults,
-	}
-
-	err = g.storeAllDocumentsToVectorStore(ctx, storeOptions)
-	if err != nil {
-		return nil, fmt.Errorf("failed to store documents to vector store: %w", err)
-	}
-
-	// Step 12: Store segment metadata (Weight, Score, Vote) to Store if configured
-	if g.Store != nil {
-		err = g.storeSegmentMetadataToStore(ctx, docID, chunks, collectionIDs.Store)
-		if err != nil {
-			g.Logger.Warnf("Failed to store segment metadata to Store: %v", err)
-		}
-	}
-
-	// Collect and return all segment IDs
-	segmentIDs := make([]string, len(chunks))
-	for i, chunk := range chunks {
-		segmentIDs[i] = chunk.ID
-	}
-
-	return segmentIDs, nil
-}
-
-// UpdateSegments updates segments manually
-func (g *GraphRag) UpdateSegments(ctx context.Context, segmentTexts []types.SegmentText, options *types.UpsertOptions) (int, error) {
-	// Step 1: Validate input - all segments must have IDs
-	for i, segmentText := range segmentTexts {
-		if segmentText.ID == "" {
-			return 0, fmt.Errorf("segment at index %d is missing ID - all segments must have IDs for update", i)
-		}
-	}
-
-	if len(segmentTexts) == 0 {
-		return 0, nil
-	}
-
-	// Step 2: Get existing chunks to read doc_id and validate segments exist
-	segmentIDs := make([]string, 0, len(segmentTexts))
-	for _, segment := range segmentTexts {
-		segmentIDs = append(segmentIDs, segment.ID)
-	}
-
-	// docID, collectionID, err := g.getDocIDFromExistingSegments(ctx, segmentIDs)
-	// if err != nil {
-	// 	return 0, fmt.Errorf("failed to get doc_id from existing segments: %w", err)
-	// }
-
-	// Step 3: Get collection ID and doc ID from options (for update segments)
-	collectionID := options.CollectionID
-	docID := options.DocID
-
-	// Step 3: Prepare options by copying and setting necessary fields
-	opts := &types.UpsertOptions{}
-	if options != nil {
-		*opts = *options
-	}
-	opts.CollectionID = collectionID
-	opts.DocID = docID
-
-	// Step 4: Get collection IDs for vector and graph storage
-	collectionIDs, err := utils.GetCollectionIDs(collectionID)
-	if err != nil {
-		return 0, fmt.Errorf("failed to generate collection IDs: %w", err)
-	}
-
-	// Step 5: Convert SegmentTexts to Chunks for processing
-	chunks, err := g.convertSegmentTextsToChunks(segmentTexts, docID)
-	if err != nil {
-		return 0, fmt.Errorf("failed to convert segment texts to chunks: %w", err)
-	}
-
-	// Step 6: Create the callback for progress tracking
-	var embeddingTexts []string = []string{}
-	var embeddingIndexesMap map[*types.Chunk]int = map[*types.Chunk]int{}
-	var cb = MakeUpsertCallback(docID, nil, opts.Progress)
-
-	// Step 7: Require embedding and extraction configurations
-	if opts.Embedding == nil {
-		return 0, fmt.Errorf("embedding configuration is required for UpdateSegments operation")
-	}
-
-	if g.Graph != nil && opts.Extraction == nil {
-		return 0, fmt.Errorf("extraction configuration is required when graph store is configured")
-	}
-
-	// Step 8: If Graph is configured, remove existing entities and relationships for these segments
-	if g.Graph != nil && opts.Extraction != nil {
-		err = g.removeEntitiesAndRelationshipsForSegments(ctx, collectionIDs.Graph, segmentIDs)
-		if err != nil {
-			g.Logger.Warnf("Failed to remove existing entities and relationships: %v", err)
-		}
-	}
-
-	// Step 9: Store embedding indexes for chunks
-	for _, chunk := range chunks {
-		embeddingTexts = append(embeddingTexts, chunk.Text)
-		embeddingIndexesMap[chunk] = len(embeddingTexts) - 1
-	}
-
-	// Step 10: Extract entities and relationships from chunks (if Graph is configured)
-	allEntities, allRelationships, entityIndexMap, relationshipIndexMap, err := g.extractEntitiesAndRelationships(ctx, chunks, opts, cb, &embeddingTexts)
-	if err != nil {
-		return 0, fmt.Errorf("failed to extract entities and relationships: %w", err)
-	}
-
-	// Step 11: Embed all texts (chunks + entities + relationships)
-	embeddings, err := opts.Embedding.EmbedDocuments(ctx, embeddingTexts, cb.Embedding)
-	if err != nil {
-		return 0, fmt.Errorf("failed to embed the documents: %w", err)
-	}
-
-	// Step 12: Store entities and relationships to graph store (if available)
-	var actualEntityIDs []string
-	var actualRelationshipIDs []string
-	var entityIDMap = make(map[string]string)
-	var relationshipIDMap = make(map[string]string)
-
-	var entityDeduplicationResults map[string]*EntityDeduplicationResult
-	var relationshipDeduplicationResults map[string]*RelationshipDeduplicationResult
-
-	if g.Graph != nil && opts.Extraction != nil && (len(allEntities) > 0 || len(allRelationships) > 0) {
-		// Store entities to graph store
-		if len(allEntities) > 0 {
-			actualEntityIDs, entityDeduplicationResults, err = g.storeEntitiesToGraphStore(ctx, allEntities, collectionIDs.Graph, docID)
-			if err != nil {
-				return 0, fmt.Errorf("failed to store entities to graph store: %w", err)
-			}
-
-			// Create mapping from original IDs to actual IDs
-			for i, entity := range allEntities {
-				if i < len(actualEntityIDs) {
-					entityIDMap[entity.ID] = actualEntityIDs[i]
-				}
-			}
-		}
-
-		// Store relationships to graph store
-		if len(allRelationships) > 0 {
-			actualRelationshipIDs, relationshipDeduplicationResults, err = g.storeRelationshipsToGraphStore(ctx, allRelationships, collectionIDs.Graph, docID)
-			if err != nil {
-				return 0, fmt.Errorf("failed to store relationships to graph store: %w", err)
-			}
-
-			// Create mapping from original IDs to actual IDs
-			for i, relationship := range allRelationships {
-				if i < len(actualRelationshipIDs) {
-					relationshipIDMap[relationship.ID] = actualRelationshipIDs[i]
-				}
-			}
-		}
-
-		// Update chunks with actual IDs from graph database
-		g.updateChunksWithActualIds(chunks, entityIDMap, relationshipIDMap)
-	}
-
-	// Step 13: Update all documents in vector store (chunks + entities + relationships)
-	storeOptions := &StoreDocumentsOptions{
-		Chunks:                           chunks,
-		Entities:                         allEntities,
-		Relationships:                    allRelationships,
-		Embeddings:                       embeddings,
-		EmbeddingIndexesMap:              embeddingIndexesMap,
-		EntityIndexMap:                   entityIndexMap,
-		RelationshipIndexMap:             relationshipIndexMap,
-		SourceFile:                       "",
-		ConvertMetadata:                  make(map[string]interface{}),
-		UserMetadata:                     opts.Metadata,
-		VectorCollectionName:             collectionIDs.Vector,
-		CollectionID:                     collectionID,
-		DocID:                            docID,
-		EntityDeduplicationResults:       entityDeduplicationResults,
-		RelationshipDeduplicationResults: relationshipDeduplicationResults,
-	}
-
-	err = g.storeAllDocumentsToVectorStore(ctx, storeOptions)
-	if err != nil {
-		return 0, fmt.Errorf("failed to update documents in vector store: %w", err)
-	}
-
-	// Step 14: Update segment metadata in Store if configured and user metadata contains vote/weight/score
-	if g.Store != nil && opts.Metadata != nil {
-		err = g.updateSegmentMetadataInStore(ctx, docID, segmentTexts, opts.Metadata, collectionIDs.Store)
-		if err != nil {
-			g.Logger.Warnf("Failed to update segment metadata in Store: %v", err)
-		}
-	}
-
-	return len(segmentTexts), nil
-}
-
-// RemoveSegments removes segments by IDs
-func (g *GraphRag) RemoveSegments(ctx context.Context, docID string, segmentIDs []string) (int, error) {
-	if len(segmentIDs) == 0 {
-		return 0, nil
-	}
-
-	if docID == "" {
-		return 0, fmt.Errorf("docID cannot be empty")
-	}
-
-	g.Logger.Infof("Starting to remove %d segments for document %s", len(segmentIDs), docID)
-
-	// Extract GraphName from docID
-	graphName, _ := utils.ExtractCollectionIDFromDocID(docID)
-	if graphName == "" {
-		graphName = "default"
-	}
-
-	// Step 2: Get collection IDs for this graph
-	collectionIDs, err := utils.GetCollectionIDs(graphName)
-	if err != nil {
-		return 0, fmt.Errorf("failed to get collection IDs for graph %s: %w", graphName, err)
-	}
-
-	// Step 3: Remove segments from vector store
-	if g.Vector != nil {
-		err := g.removeSegmentsFromVectorStore(ctx, collectionIDs.Vector, segmentIDs)
-		if err != nil {
-			g.Logger.Warnf("Failed to remove segments from vector store %s: %v", collectionIDs.Vector, err)
-		} else {
-			g.Logger.Infof("Removed segments from vector store: %s", collectionIDs.Vector)
-		}
-	}
-
-	// Step 4: Update entities and relationships in graph store
-	if g.Graph != nil && g.Graph.IsConnected() {
-		err := g.removeEntitiesAndRelationshipsForSegments(ctx, collectionIDs.Graph, segmentIDs)
-		if err != nil {
-			g.Logger.Warnf("Failed to update entities/relationships in graph store %s: %v", collectionIDs.Graph, err)
-		} else {
-			g.Logger.Infof("Updated entities/relationships in graph store: %s", collectionIDs.Graph)
-		}
-	}
-
-	// Step 5: Remove segment metadata from Store
-	if g.Store != nil {
-		g.removeSegmentsFromStore(ctx, docID, segmentIDs)
-	}
-
-	removedCount := len(segmentIDs)
-	g.Logger.Infof("Successfully removed %d segments", removedCount)
-	return removedCount, nil
-}
-
-// RemoveSegmentsByDocID removes all segments of a document
-func (g *GraphRag) RemoveSegmentsByDocID(ctx context.Context, docID string) (int, error) {
-	if docID == "" {
-		return 0, fmt.Errorf("docID cannot be empty")
-	}
-
-	g.Logger.Infof("Starting to remove all segments for document: %s", docID)
-
-	// Step 1: Parse GraphName from docID
-	graphName, _ := utils.ExtractCollectionIDFromDocID(docID)
-	if graphName == "" {
-		graphName = "default"
-	}
-
-	// Step 2: Get collection IDs for this graph
-	collectionIDs, err := utils.GetCollectionIDs(graphName)
-	if err != nil {
-		return 0, fmt.Errorf("failed to get collection IDs for graph %s: %w", graphName, err)
-	}
-
-	// Step 3: Count segments before deletion (for return value)
-	segmentCount := 0
-	if g.Vector != nil {
-		exists, err := g.Vector.CollectionExists(ctx, collectionIDs.Vector)
-		if err == nil && exists {
-			// Get count of segments for this document
-			listOpts := &types.ListDocumentsOptions{
-				CollectionName: collectionIDs.Vector,
-				Filter: map[string]interface{}{
-					"doc_id":        docID,
-					"document_type": "chunk",
-				},
-				Limit:          1000,
-				IncludeVector:  false,
-				IncludePayload: false,
-			}
-			result, err := g.Vector.ListDocuments(ctx, listOpts)
-			if err == nil {
-				segmentCount = len(result.Documents)
-			}
-		}
-	}
-
-	// Step 4: Remove segments from vector store by doc_id filter
-	if g.Vector != nil {
-		err := g.removeSegmentsByDocIDFromVectorStore(ctx, collectionIDs.Vector, docID)
-		if err != nil {
-			g.Logger.Warnf("Failed to remove segments from vector store %s: %v", collectionIDs.Vector, err)
-		} else {
-			g.Logger.Infof("Removed segments from vector store: %s", collectionIDs.Vector)
-		}
-	}
-
-	// Step 5: Process entities and relationships in graph store (same as document deletion)
-	if g.Graph != nil && g.Graph.IsConnected() {
-		err := g.removeDocsFromGraphStore(ctx, collectionIDs.Graph, []string{docID})
-		if err != nil {
-			g.Logger.Warnf("Failed to process entities/relationships in graph store %s: %v", collectionIDs.Graph, err)
-		} else {
-			g.Logger.Infof("Processed entities/relationships in graph store: %s", collectionIDs.Graph)
-		}
-	}
-
-	// Step 6: Remove all segment metadata from Store for this docID
-	if g.Store != nil {
-		g.removeAllSegmentMetadataFromStore(ctx, docID)
-	}
-
-	g.Logger.Infof("Successfully removed %d segments for document: %s", segmentCount, docID)
-	return segmentCount, nil
-}
-
-// ================================================================================================
-// Query Operations - Exported Methods
-// ================================================================================================
-
-// GetSegment gets a single segment by ID
-func (g *GraphRag) GetSegment(ctx context.Context, docID string, segmentID string) (*types.Segment, error) {
-	if docID == "" {
-		return nil, fmt.Errorf("docID cannot be empty")
-	}
-
-	if segmentID == "" {
-		return nil, fmt.Errorf("segmentID cannot be empty")
-	}
-
-	g.Logger.Debugf("Getting segment by ID: %s for document: %s", segmentID, docID)
-
-	// Get segments using GetSegments
-	segments, err := g.GetSegments(ctx, docID, []string{segmentID})
-	if err != nil {
-		return nil, fmt.Errorf("failed to get segment: %w", err)
-	}
-
-	if len(segments) == 0 {
-		return nil, fmt.Errorf("segment %s not found", segmentID)
-	}
-
-	return &segments[0], nil
-}
-
-// GetSegmentParents gets parent tree of a given segment
-func (g *GraphRag) GetSegmentParents(ctx context.Context, docID string, segmentID string) (*types.SegmentTree, error) {
-	if docID == "" {
-		return nil, fmt.Errorf("docID cannot be empty")
-	}
-
-	if segmentID == "" {
-		return nil, fmt.Errorf("segmentID cannot be empty")
-	}
-
-	g.Logger.Debugf("Getting parent tree for segment ID: %s in document: %s", segmentID, docID)
-
-	// Get the target segment first
-	segment, err := g.GetSegment(ctx, docID, segmentID)
-	if err != nil {
-		return nil, fmt.Errorf("failed to get segment: %w", err)
-	}
-
-	// Build the tree structure starting from the target segment
-	tree, err := g.buildSegmentTree(ctx, docID, segment)
-	if err != nil {
-		return nil, fmt.Errorf("failed to build segment tree: %w", err)
-	}
-
-	g.Logger.Debugf("Successfully built parent tree for segment %s", segmentID)
-	return tree, nil
-}
-
-// GetSegments gets segments by IDs
-func (g *GraphRag) GetSegments(ctx context.Context, docID string, segmentIDs []string) ([]types.Segment, error) {
-	if len(segmentIDs) == 0 {
-		return []types.Segment{}, nil
-	}
-
-	if docID == "" {
-		return nil, fmt.Errorf("docID cannot be empty")
-	}
-
-	g.Logger.Debugf("Getting %d segments by IDs for document %s", len(segmentIDs), docID)
-
-	// Extract GraphName from docID
-	graphName, _ := utils.ExtractCollectionIDFromDocID(docID)
-	if graphName == "" {
-		graphName = "default"
-	}
-
-	// Query segment data from all configured databases
-	segmentData, err := g.querySegmentData(ctx, &segmentQueryOptions{
-		GraphName:  graphName,
-		DocID:      docID,
-		SegmentIDs: segmentIDs,
-		QueryType:  "by_ids",
-	})
-	if err != nil {
-		return nil, fmt.Errorf("failed to query segment data: %w", err)
-	}
-
-	// Assemble segments from the queried data
-	segments := g.assembleSegments(segmentData, graphName, docID)
-
-	g.Logger.Debugf("Successfully retrieved %d segments", len(segments))
-	return segments, nil
-}
-
-// ListSegments lists segments of a document with pagination (deprecated)
-func (g *GraphRag) ListSegments(ctx context.Context, docID string, options *types.ListSegmentsOptions) (*types.PaginatedSegmentsResult, error) {
-	if docID == "" {
-		return nil, fmt.Errorf("docID cannot be empty")
-	}
-
-	g.Logger.Debugf("Listing segments for document: %s", docID)
-
-	// Parse GraphName from docID
-	graphName, _ := utils.ExtractCollectionIDFromDocID(docID)
-	if graphName == "" {
-		graphName = "default"
-	}
-
-	// Set default options
-	if options == nil {
-		options = &types.ListSegmentsOptions{}
-	}
-
-	// Set default limit
-	limit := options.Limit
-	if limit <= 0 {
-		limit = 100 // Default limit
-	}
-
-	// Query segment data from all configured databases with pagination
-	segmentData, total, hasMore, nextOffset, err := g.querySegmentDataWithPagination(ctx, &segmentQueryOptions{
-		GraphName:            graphName,
-		DocID:                docID,
-		QueryType:            "list",
-		Limit:                limit,
-		Offset:               options.Offset,
-		Filter:               options.Filter,
-		OrderBy:              options.OrderBy,
-		Fields:               options.Fields,
-		IncludeNodes:         options.IncludeNodes,
-		IncludeRelationships: options.IncludeRelationships,
-		IncludeMetadata:      options.IncludeMetadata,
-	})
-	if err != nil {
-		return nil, fmt.Errorf("failed to query segment data: %w", err)
-	}
-
-	// Assemble segments from the queried data
-	segments := g.assembleSegments(segmentData, graphName, docID)
-
-	result := &types.PaginatedSegmentsResult{
-		Segments:   segments,
-		Total:      total,
-		HasMore:    hasMore,
-		NextOffset: nextOffset,
-	}
-
-	g.Logger.Debugf("Successfully listed %d segments for document %s", len(segments), docID)
-	return result, nil
-}
-
-// ScrollSegments scrolls through segments of a document with iterator-style pagination
-func (g *GraphRag) ScrollSegments(ctx context.Context, docID string, options *types.ScrollSegmentsOptions) (*types.SegmentScrollResult, error) {
-	if docID == "" {
-		return nil, fmt.Errorf("docID cannot be empty")
-	}
-
-	g.Logger.Debugf("Scrolling segments for document: %s", docID)
-
-	// Parse GraphName from docID
-	graphName, _ := utils.ExtractCollectionIDFromDocID(docID)
-	if graphName == "" {
-		graphName = "default"
-	}
-
-	// Set default options
-	if options == nil {
-		options = &types.ScrollSegmentsOptions{}
-	}
-
-	// Set default limit
-	limit := options.Limit
-	if limit <= 0 {
-		limit = 100 // Default limit
-	}
-
-	// Query segment data from all configured databases with scroll
-	segmentData, scrollID, hasMore, err := g.querySegmentDataWithScroll(ctx, &segmentQueryOptions{
-		GraphName:            graphName,
-		DocID:                docID,
-		QueryType:            "scroll",
-		BatchSize:            limit,
-		ScrollID:             options.ScrollID,
-		Filter:               options.Filter,
-		OrderBy:              options.OrderBy,
-		Fields:               options.Fields,
-		IncludeNodes:         options.IncludeNodes,
-		IncludeRelationships: options.IncludeRelationships,
-		IncludeMetadata:      options.IncludeMetadata,
-	})
-	if err != nil {
-		return nil, fmt.Errorf("failed to query segment data: %w", err)
-	}
-
-	// Assemble segments from the queried data
-	segments := g.assembleSegments(segmentData, graphName, docID)
-
-	result := &types.SegmentScrollResult{
-		Segments: segments,
-		ScrollID: scrollID,
-		HasMore:  hasMore,
-	}
-
-	g.Logger.Debugf("Successfully scrolled %d segments for document %s", len(segments), docID)
-	return result, nil
-}
-
-// ================================================================================================
-// Internal Helper Methods - Conversion and Processing
-// ================================================================================================
-
-// buildSegmentTree recursively builds a tree structure of segment parents using metadata
-func (g *GraphRag) buildSegmentTree(ctx context.Context, docID string, segment *types.Segment) (*types.SegmentTree, error) {
-	// Extract depth from segment metadata
-	segmentDepth := g.extractDepthFromSegment(segment)
-
-	// Create the tree node for the current segment
-	tree := &types.SegmentTree{
-		Segment: segment,
-		Parent:  nil,
-		Depth:   segmentDepth,
-	}
-
-	// Extract parent ID from segment metadata (only one parent in document hierarchy)
-	parentID := g.extractParentIDFromSegment(segment)
-	if parentID == "" {
-		// No parent found, this is a root node
-		return tree, nil
-	}
-
-	// Get the parent segment
-	parentSegment, err := g.GetSegment(ctx, docID, parentID)
-	if err != nil {
-		g.Logger.Warnf("Failed to get parent segment %s: %v", parentID, err)
-		// Return the tree without parent rather than failing completely
-		return tree, nil
-	}
-
-	// Build parent tree recursively
-	parentTree, err := g.buildSegmentTree(ctx, docID, parentSegment)
-	if err != nil {
-		g.Logger.Warnf("Failed to build parent tree for segment %s: %v", parentSegment.ID, err)
-		return tree, nil
-	}
-
-	// Set the parent
-	tree.Parent = parentTree
-
-	return tree, nil
-}
-
-// extractParentIDFromSegment extracts the parent ID from segment metadata
-func (g *GraphRag) extractParentIDFromSegment(segment *types.Segment) string {
-	if segment == nil {
-		return ""
-	}
-
-	// Strategy 1: Check segment's Parents field (take the first one, should only be one)
-	if len(segment.Parents) > 0 {
-		return segment.Parents[0]
-	}
-
-	// Strategy 2: Check chunk_details in metadata for parent_id
-	if segment.Metadata != nil {
-		if chunkDetails, ok := segment.Metadata["chunk_details"].(map[string]interface{}); ok {
-			if parentID := types.SafeExtractString(chunkDetails["parent_id"], ""); parentID != "" {
-				return parentID
-			}
-		}
-
-		// Strategy 3: Check direct parent_id in metadata
-		if parentID := types.SafeExtractString(segment.Metadata["parent_id"], ""); parentID != "" {
-			return parentID
-		}
-	}
-
-	return ""
-}
-
-// extractDepthFromSegment extracts depth value from segment metadata
-func (g *GraphRag) extractDepthFromSegment(segment *types.Segment) int {
-	if segment == nil {
-		return 0
-	}
-
-	// Strategy 1: Check chunk_details in metadata for depth
-	if segment.Metadata != nil {
-		if chunkDetails, ok := segment.Metadata["chunk_details"].(map[string]interface{}); ok {
-			if depth := types.SafeExtractInt(chunkDetails["depth"], 0); depth > 0 {
-				return depth
-			}
-		}
-
-		// Strategy 2: Check direct depth in metadata
-		if depth := types.SafeExtractInt(segment.Metadata["depth"], 0); depth > 0 {
-			return depth
-		}
-	}
-
-	// Strategy 3: Fallback - calculate from parent chain length if available
-	// This is a fallback when depth is not stored in metadata
-	if len(segment.Parents) > 0 {
-		// If we have parents, we're at least depth 1 (not root)
-		// This is a rough estimation, the actual depth should be in metadata
-		return len(segment.Parents)
-	}
-
-	// Default: assume root level (depth 1) if no depth information found
-	return 1
-}
 
 // convertSegmentTextsToChunks converts SegmentTexts to Chunks for processing
 func (g *GraphRag) convertSegmentTextsToChunks(segmentTexts []types.SegmentText, docID string) ([]*types.Chunk, error) {
@@ -868,196 +119,8 @@ func (g *GraphRag) applyMetadataToChunk(chunk *types.Chunk, metadata map[string]
 }
 
 // ================================================================================================
-// Internal Helper Methods - Vector Store Operations
+// Exported Utility Functions - Store Operations
 // ================================================================================================
-
-// removeSegmentsFromVectorStore removes segments from vector database
-func (g *GraphRag) removeSegmentsFromVectorStore(ctx context.Context, collectionName string, segmentIDs []string) error {
-	// Check if collection exists
-	exists, err := g.Vector.CollectionExists(ctx, collectionName)
-	if err != nil {
-		return fmt.Errorf("failed to check collection existence: %w", err)
-	}
-	if !exists {
-		g.Logger.Infof("Vector collection %s does not exist, skipping vector deletion", collectionName)
-		return nil
-	}
-
-	// Delete segments (chunks) by IDs
-	err = g.Vector.DeleteDocuments(ctx, &types.DeleteDocumentOptions{
-		CollectionName: collectionName,
-		IDs:            segmentIDs,
-	})
-	if err != nil {
-		return fmt.Errorf("failed to delete segments: %w", err)
-	}
-
-	g.Logger.Debugf("Deleted %d segments from vector store", len(segmentIDs))
-	return nil
-}
-
-// removeSegmentsByDocIDFromVectorStore removes all segments for a document from vector store using filter
-func (g *GraphRag) removeSegmentsByDocIDFromVectorStore(ctx context.Context, collectionName string, docID string) error {
-	// Check if collection exists
-	exists, err := g.Vector.CollectionExists(ctx, collectionName)
-	if err != nil {
-		return fmt.Errorf("failed to check collection existence: %w", err)
-	}
-	if !exists {
-		g.Logger.Infof("Vector collection %s does not exist, skipping vector deletion", collectionName)
-		return nil
-	}
-
-	// Delete all chunks (segments) for this document
-	chunksFilter := map[string]interface{}{
-		"doc_id":        docID,
-		"document_type": "chunk",
-	}
-
-	err = g.Vector.DeleteDocuments(ctx, &types.DeleteDocumentOptions{
-		CollectionName: collectionName,
-		Filter:         chunksFilter,
-	})
-	if err != nil {
-		return fmt.Errorf("failed to delete segments for document %s: %w", docID, err)
-	}
-
-	g.Logger.Debugf("Deleted all segments for document %s from vector store", docID)
-	return nil
-}
-
-// findSegmentsByDocID finds all segment IDs for a given document from vector store
-func (g *GraphRag) findSegmentsByDocID(ctx context.Context, collectionName string, docID string) ([]string, error) {
-	// Check if collection exists
-	exists, err := g.Vector.CollectionExists(ctx, collectionName)
-	if err != nil {
-		return nil, fmt.Errorf("failed to check collection existence: %w", err)
-	}
-	if !exists {
-		g.Logger.Infof("Vector collection %s does not exist, no segments to find", collectionName)
-		return []string{}, nil
-	}
-
-	// Use ListDocuments to find all chunks (segments) for this document
-	// We filter by doc_id and document_type = "chunk"
-	listOpts := &types.ListDocumentsOptions{
-		CollectionName: collectionName,
-		Filter: map[string]interface{}{
-			"doc_id":        docID,
-			"document_type": "chunk",
-		},
-		Limit:          1000, // Set a reasonable limit to avoid too much data
-		IncludeVector:  false,
-		IncludePayload: false, // We only need the IDs
-	}
-
-	result, err := g.Vector.ListDocuments(ctx, listOpts)
-	if err != nil {
-		return nil, fmt.Errorf("failed to list segments: %w", err)
-	}
-
-	// Extract segment IDs from the result
-	segmentIDs := make([]string, 0, len(result.Documents))
-	for _, doc := range result.Documents {
-		if doc.ID != "" {
-			segmentIDs = append(segmentIDs, doc.ID)
-		}
-	}
-
-	g.Logger.Debugf("Found %d segments for document %s", len(segmentIDs), docID)
-	return segmentIDs, nil
-}
-
-// ================================================================================================
-// Internal Helper Methods - Store Operations
-// ================================================================================================
-
-// removeSegmentsFromStore removes segment metadata from Store
-func (g *GraphRag) removeSegmentsFromStore(ctx context.Context, docID string, segmentIDs []string) {
-	for _, segmentID := range segmentIDs {
-		// Delete Weight
-		err := g.deleteSegmentValue(docID, segmentID, StoreKeyWeight)
-		if err != nil {
-			g.Logger.Warnf("Failed to delete weight for segment %s: %v", segmentID, err)
-		}
-
-		// Delete Score
-		err = g.deleteSegmentValue(docID, segmentID, StoreKeyScore)
-		if err != nil {
-			g.Logger.Warnf("Failed to delete score for segment %s: %v", segmentID, err)
-		}
-
-		// Delete Vote
-		err = g.deleteSegmentValue(docID, segmentID, StoreKeyVote)
-		if err != nil {
-			g.Logger.Warnf("Failed to delete vote for segment %s: %v", segmentID, err)
-		}
-	}
-}
-
-// removeAllSegmentMetadataFromStore removes all segment metadata for a document from Store
-func (g *GraphRag) removeAllSegmentMetadataFromStore(ctx context.Context, docID string) {
-	if g.Store == nil {
-		return
-	}
-
-	g.Logger.Debugf("Attempting to remove segment metadata for document %s", docID)
-
-	// Parse CollectionID from docID to find the right collection
-	collectionID, _ := utils.ExtractCollectionIDFromDocID(docID)
-	if collectionID == "" {
-		collectionID = "default"
-	}
-
-	// Get collection IDs for this collection
-	collectionIDs, err := utils.GetCollectionIDs(collectionID)
-	if err != nil {
-		g.Logger.Warnf("Failed to get collection IDs for document %s: %v", docID, err)
-		return
-	}
-
-	// Find all segment IDs for this document from vector store
-	segmentIDs, err := g.findSegmentsByDocID(ctx, collectionIDs.Vector, docID)
-	if err != nil {
-		g.Logger.Warnf("Failed to find segments for document %s: %v", docID, err)
-		return
-	}
-
-	if len(segmentIDs) == 0 {
-		g.Logger.Debugf("No segments found for document %s, no Store cleanup needed", docID)
-		return
-	}
-
-	// Delete Store metadata for each segment using the helper function
-	removedCount := 0
-	for _, segmentID := range segmentIDs {
-		// Delete Weight
-		err := g.deleteSegmentValue(docID, segmentID, StoreKeyWeight)
-		if err != nil {
-			g.Logger.Warnf("Failed to delete weight for segment %s: %v", segmentID, err)
-		} else {
-			removedCount++
-		}
-
-		// Delete Score
-		err = g.deleteSegmentValue(docID, segmentID, StoreKeyScore)
-		if err != nil {
-			g.Logger.Warnf("Failed to delete score for segment %s: %v", segmentID, err)
-		} else {
-			removedCount++
-		}
-
-		// Delete Vote
-		err = g.deleteSegmentValue(docID, segmentID, StoreKeyVote)
-		if err != nil {
-			g.Logger.Warnf("Failed to delete vote for segment %s: %v", segmentID, err)
-		} else {
-			removedCount++
-		}
-	}
-
-	g.Logger.Infof("Segment metadata cleanup completed for document %s: removed %d Store entries for %d segments", docID, removedCount, len(segmentIDs))
-}
 
 // storeSegmentMetadataToStore stores segment metadata (Weight, Score, Vote) to Store and/or Vector DB
 func (g *GraphRag) storeSegmentMetadataToStore(ctx context.Context, docID string, chunks []*types.Chunk, storeCollectionName string) error {
@@ -1132,178 +195,7 @@ func (g *GraphRag) updateSegmentMetadataInStore(ctx context.Context, docID strin
 }
 
 // ================================================================================================
-// Internal Helper Methods - Graph Operations
-// ================================================================================================
-
-// removeEntitiesAndRelationshipsForSegments removes entities and relationships associated with specific segments
-func (g *GraphRag) removeEntitiesAndRelationshipsForSegments(ctx context.Context, graphName string, segmentIDs []string) error {
-	if len(segmentIDs) == 0 {
-		return nil
-	}
-
-	// Check if graph exists
-	exists, err := g.Graph.GraphExists(ctx, graphName)
-	if err != nil {
-		return fmt.Errorf("failed to check graph existence: %w", err)
-	}
-	if !exists {
-		g.Logger.Infof("Graph %s does not exist, skipping entity/relationship removal", graphName)
-		return nil
-	}
-
-	// Remove entities and relationships that were extracted from these specific chunks
-	// We need to query entities and relationships that have these chunk IDs in their source_chunks
-
-	// Remove entities associated with these segments
-	for _, segmentID := range segmentIDs {
-		// Query entities that have this segment in their source_chunks
-		queryOpts := &types.GraphQueryOptions{
-			GraphName: graphName,
-			QueryType: "cypher",
-			Query:     "MATCH (n) WHERE $segmentID IN n.source_chunks RETURN n",
-			Parameters: map[string]interface{}{
-				"segmentID": segmentID,
-			},
-		}
-
-		result, err := g.Graph.Query(ctx, queryOpts)
-		if err != nil {
-			g.Logger.Warnf("Failed to query entities for segment %s: %v", segmentID, err)
-			continue
-		}
-
-		// Remove segmentID from source_chunks, delete entity if source_chunks becomes empty
-		for _, node := range result.Nodes {
-			sourceChunksInterface, exists := node.Properties["source_chunks"]
-			if !exists {
-				continue
-			}
-
-			var newSourceChunks []string
-			switch sourceChunksValue := sourceChunksInterface.(type) {
-			case []interface{}:
-				for _, chunk := range sourceChunksValue {
-					if chunkStr, ok := chunk.(string); ok && chunkStr != segmentID {
-						newSourceChunks = append(newSourceChunks, chunkStr)
-					}
-				}
-			case []string:
-				for _, chunk := range sourceChunksValue {
-					if chunk != segmentID {
-						newSourceChunks = append(newSourceChunks, chunk)
-					}
-				}
-			}
-
-			// If no more source chunks, delete the entity
-			if len(newSourceChunks) == 0 {
-				err := g.Graph.DeleteNodes(ctx, &types.DeleteNodesOptions{
-					GraphName:  graphName,
-					IDs:        []string{node.ID},
-					DeleteRels: true,
-				})
-				if err != nil {
-					g.Logger.Warnf("Failed to delete entity %s: %v", node.ID, err)
-				}
-			} else {
-				// Update entity's source_chunks
-				node.Properties["source_chunks"] = newSourceChunks
-				graphNode := &types.GraphNode{
-					ID:         node.ID,
-					Labels:     node.Labels,
-					Properties: node.Properties,
-				}
-
-				_, err := g.Graph.AddNodes(ctx, &types.AddNodesOptions{
-					GraphName: graphName,
-					Nodes:     []*types.GraphNode{graphNode},
-					Upsert:    true,
-				})
-				if err != nil {
-					g.Logger.Warnf("Failed to update entity %s source_chunks: %v", node.ID, err)
-				}
-			}
-		}
-	}
-
-	// Remove relationships associated with these segments
-	for _, segmentID := range segmentIDs {
-		// Query relationships that have this segment in their source_chunks
-		queryOpts := &types.GraphQueryOptions{
-			GraphName: graphName,
-			QueryType: "cypher",
-			Query:     "MATCH ()-[r]->() WHERE $segmentID IN r.source_chunks RETURN r",
-			Parameters: map[string]interface{}{
-				"segmentID": segmentID,
-			},
-		}
-
-		result, err := g.Graph.Query(ctx, queryOpts)
-		if err != nil {
-			g.Logger.Warnf("Failed to query relationships for segment %s: %v", segmentID, err)
-			continue
-		}
-
-		// Remove segmentID from source_chunks, delete relationship if source_chunks becomes empty
-		for _, rel := range result.Relationships {
-			sourceChunksInterface, exists := rel.Properties["source_chunks"]
-			if !exists {
-				continue
-			}
-
-			var newSourceChunks []string
-			switch sourceChunksValue := sourceChunksInterface.(type) {
-			case []interface{}:
-				for _, chunk := range sourceChunksValue {
-					if chunkStr, ok := chunk.(string); ok && chunkStr != segmentID {
-						newSourceChunks = append(newSourceChunks, chunkStr)
-					}
-				}
-			case []string:
-				for _, chunk := range sourceChunksValue {
-					if chunk != segmentID {
-						newSourceChunks = append(newSourceChunks, chunk)
-					}
-				}
-			}
-
-			// If no more source chunks, delete the relationship
-			if len(newSourceChunks) == 0 {
-				err := g.Graph.DeleteRelationships(ctx, &types.DeleteRelationshipsOptions{
-					GraphName: graphName,
-					IDs:       []string{rel.ID},
-				})
-				if err != nil {
-					g.Logger.Warnf("Failed to delete relationship %s: %v", rel.ID, err)
-				}
-			} else {
-				// Update relationship's source_chunks
-				rel.Properties["source_chunks"] = newSourceChunks
-				graphRel := &types.GraphRelationship{
-					ID:         rel.ID,
-					Type:       rel.Type,
-					StartNode:  rel.StartNode,
-					EndNode:    rel.EndNode,
-					Properties: rel.Properties,
-				}
-
-				_, err := g.Graph.AddRelationships(ctx, &types.AddRelationshipsOptions{
-					GraphName:     graphName,
-					Relationships: []*types.GraphRelationship{graphRel},
-					Upsert:        true,
-				})
-				if err != nil {
-					g.Logger.Warnf("Failed to update relationship %s source_chunks: %v", rel.ID, err)
-				}
-			}
-		}
-	}
-
-	return nil
-}
-
-// ================================================================================================
-// Internal Helper Methods - Query System
+// Internal Types and Structures - Query System
 // ================================================================================================
 
 // segmentQueryOptions represents options for querying segment data
@@ -1337,6 +229,10 @@ type segmentQueryResult struct {
 	Relationships []types.GraphRelationship // Relationships from graph database
 	StoreData     map[string]interface{}    // Metadata from KV store
 }
+
+// ================================================================================================
+// Exported Utility Functions - Query System
+// ================================================================================================
 
 // querySegmentData queries segment data from vector, graph, and store databases concurrently
 func (g *GraphRag) querySegmentData(ctx context.Context, opts *segmentQueryOptions) (*segmentQueryResult, error) {
@@ -1426,194 +322,6 @@ func (g *GraphRag) querySegmentData(ctx context.Context, opts *segmentQueryOptio
 	return result, nil
 }
 
-// querySegmentDataWithPagination queries segment data with pagination support
-func (g *GraphRag) querySegmentDataWithPagination(ctx context.Context, opts *segmentQueryOptions) (*segmentQueryResult, int64, bool, int, error) {
-	// Get collection IDs
-	collectionIDs, err := utils.GetCollectionIDs(opts.GraphName)
-	if err != nil {
-		return nil, 0, false, 0, fmt.Errorf("failed to get collection IDs: %w", err)
-	}
-
-	result := &segmentQueryResult{
-		StoreData: make(map[string]interface{}),
-	}
-
-	var totalCount int64 = 0
-	var hasMore bool = false
-	var nextOffset int = opts.Offset
-
-	// Create error channel for concurrent operations
-	errChan := make(chan error, 3)
-
-	// Query vector database with pagination
-	go func() {
-		defer func() {
-			if r := recover(); r != nil {
-				errChan <- fmt.Errorf("panic in vector pagination query: %v", r)
-			}
-		}()
-		chunks, total, hasMorePages, nextOff, err := g.queryChunksFromVectorWithPagination(ctx, collectionIDs.Vector, opts)
-		if err != nil {
-			errChan <- fmt.Errorf("vector pagination query failed: %w", err)
-			return
-		}
-		result.Chunks = chunks
-		totalCount = total
-		hasMore = hasMorePages
-		nextOffset = nextOff
-		errChan <- nil
-	}()
-
-	// Query graph database (if configured)
-	if g.Graph != nil && g.Graph.IsConnected() && opts.IncludeNodes {
-		go func() {
-			defer func() {
-				if r := recover(); r != nil {
-					errChan <- fmt.Errorf("panic in graph query: %v", r)
-				}
-			}()
-			nodes, relationships, err := g.queryNodesAndRelationshipsFromGraph(ctx, collectionIDs.Graph, opts)
-			if err != nil {
-				errChan <- fmt.Errorf("graph query failed: %w", err)
-				return
-			}
-			result.Nodes = nodes
-			result.Relationships = relationships
-			errChan <- nil
-		}()
-	} else {
-		// Send nil error if graph is not configured
-		go func() {
-			errChan <- nil
-		}()
-	}
-
-	// Query KV store (if configured)
-	if g.Store != nil && opts.IncludeMetadata {
-		go func() {
-			defer func() {
-				if r := recover(); r != nil {
-					errChan <- fmt.Errorf("panic in store query: %v", r)
-				}
-			}()
-			storeData, err := g.queryMetadataFromStore(ctx, opts)
-			if err != nil {
-				errChan <- fmt.Errorf("store query failed: %w", err)
-				return
-			}
-			result.StoreData = storeData
-			errChan <- nil
-		}()
-	} else {
-		// Send nil error if store is not configured
-		go func() {
-			errChan <- nil
-		}()
-	}
-
-	// Wait for all queries to complete
-	for i := 0; i < 3; i++ {
-		if err := <-errChan; err != nil {
-			return nil, 0, false, 0, err
-		}
-	}
-
-	return result, totalCount, hasMore, nextOffset, nil
-}
-
-// querySegmentDataWithScroll queries segment data with scroll support
-func (g *GraphRag) querySegmentDataWithScroll(ctx context.Context, opts *segmentQueryOptions) (*segmentQueryResult, string, bool, error) {
-	// Get collection IDs
-	collectionIDs, err := utils.GetCollectionIDs(opts.GraphName)
-	if err != nil {
-		return nil, "", false, fmt.Errorf("failed to get collection IDs: %w", err)
-	}
-
-	result := &segmentQueryResult{
-		StoreData: make(map[string]interface{}),
-	}
-
-	var scrollID string = ""
-	var hasMore bool = false
-
-	// Create error channel for concurrent operations
-	errChan := make(chan error, 3)
-
-	// Query vector database with scroll
-	go func() {
-		defer func() {
-			if r := recover(); r != nil {
-				errChan <- fmt.Errorf("panic in vector scroll query: %v", r)
-			}
-		}()
-		chunks, nextScrollID, hasMorePages, err := g.queryChunksFromVectorWithScroll(ctx, collectionIDs.Vector, opts)
-		if err != nil {
-			errChan <- fmt.Errorf("vector scroll query failed: %w", err)
-			return
-		}
-		result.Chunks = chunks
-		scrollID = nextScrollID
-		hasMore = hasMorePages
-		errChan <- nil
-	}()
-
-	// Query graph database (if configured)
-	if g.Graph != nil && g.Graph.IsConnected() && opts.IncludeNodes {
-		go func() {
-			defer func() {
-				if r := recover(); r != nil {
-					errChan <- fmt.Errorf("panic in graph query: %v", r)
-				}
-			}()
-			nodes, relationships, err := g.queryNodesAndRelationshipsFromGraph(ctx, collectionIDs.Graph, opts)
-			if err != nil {
-				errChan <- fmt.Errorf("graph query failed: %w", err)
-				return
-			}
-			result.Nodes = nodes
-			result.Relationships = relationships
-			errChan <- nil
-		}()
-	} else {
-		// Send nil error if graph is not configured
-		go func() {
-			errChan <- nil
-		}()
-	}
-
-	// Query KV store (if configured)
-	if g.Store != nil && opts.IncludeMetadata {
-		go func() {
-			defer func() {
-				if r := recover(); r != nil {
-					errChan <- fmt.Errorf("panic in store query: %v", r)
-				}
-			}()
-			storeData, err := g.queryMetadataFromStore(ctx, opts)
-			if err != nil {
-				errChan <- fmt.Errorf("store query failed: %w", err)
-				return
-			}
-			result.StoreData = storeData
-			errChan <- nil
-		}()
-	} else {
-		// Send nil error if store is not configured
-		go func() {
-			errChan <- nil
-		}()
-	}
-
-	// Wait for all queries to complete
-	for i := 0; i < 3; i++ {
-		if err := <-errChan; err != nil {
-			return nil, "", false, err
-		}
-	}
-
-	return result, scrollID, hasMore, nil
-}
-
 // queryChunksFromVector queries chunks from vector database
 func (g *GraphRag) queryChunksFromVector(ctx context.Context, collectionName string, opts *segmentQueryOptions) ([]*types.Document, error) {
 	// Check if collection exists
@@ -1672,130 +380,6 @@ func (g *GraphRag) queryChunksFromVector(ctx context.Context, collectionName str
 	}
 
 	return validChunks, nil
-}
-
-// queryChunksFromVectorWithPagination queries chunks from vector database with pagination
-func (g *GraphRag) queryChunksFromVectorWithPagination(ctx context.Context, collectionName string, opts *segmentQueryOptions) ([]*types.Document, int64, bool, int, error) {
-	// Check if collection exists
-	exists, err := g.Vector.CollectionExists(ctx, collectionName)
-	if err != nil {
-		return nil, 0, false, 0, fmt.Errorf("failed to check collection existence: %w", err)
-	}
-	if !exists {
-		g.Logger.Infof("Vector collection %s does not exist, returning empty chunks", collectionName)
-		return []*types.Document{}, 0, false, 0, nil
-	}
-
-	// Build filter for documents
-	filter := map[string]interface{}{
-		"doc_id":        opts.DocID,
-		"document_type": "chunk",
-	}
-
-	// Add user-provided filter if any
-	if opts.Filter != nil {
-		for k, v := range opts.Filter {
-			filter[k] = v
-		}
-	}
-
-	// Set limit, using default if not provided
-	limit := opts.Limit
-	if limit <= 0 {
-		limit = 100
-	}
-
-	// List documents with pagination
-	listOpts := &types.ListDocumentsOptions{
-		CollectionName: collectionName,
-		Filter:         filter,
-		Limit:          limit,
-		Offset:         opts.Offset,
-		OrderBy:        opts.OrderBy,
-		Fields:         opts.Fields,
-		IncludeVector:  false,
-		IncludePayload: true,
-	}
-
-	result, err := g.Vector.ListDocuments(ctx, listOpts)
-	if err != nil {
-		return nil, 0, false, 0, fmt.Errorf("failed to list documents with pagination: %w", err)
-	}
-
-	// Filter out nil chunks
-	var validChunks []*types.Document
-	for _, chunk := range result.Documents {
-		if chunk != nil {
-			validChunks = append(validChunks, chunk)
-		}
-	}
-
-	// Calculate next offset
-	nextOffset := opts.Offset + len(validChunks)
-	if nextOffset >= int(result.Total) {
-		nextOffset = int(result.Total)
-	}
-
-	return validChunks, result.Total, result.HasMore, nextOffset, nil
-}
-
-// queryChunksFromVectorWithScroll queries chunks from vector database with scroll
-func (g *GraphRag) queryChunksFromVectorWithScroll(ctx context.Context, collectionName string, opts *segmentQueryOptions) ([]*types.Document, string, bool, error) {
-	// Check if collection exists
-	exists, err := g.Vector.CollectionExists(ctx, collectionName)
-	if err != nil {
-		return nil, "", false, fmt.Errorf("failed to check collection existence: %w", err)
-	}
-	if !exists {
-		g.Logger.Infof("Vector collection %s does not exist, returning empty chunks", collectionName)
-		return []*types.Document{}, "", false, nil
-	}
-
-	// Build filter for documents
-	filter := map[string]interface{}{
-		"doc_id":        opts.DocID,
-		"document_type": "chunk",
-	}
-
-	// Add user-provided filter if any
-	if opts.Filter != nil {
-		for k, v := range opts.Filter {
-			filter[k] = v
-		}
-	}
-
-	// Set limit, using default if not provided
-	limit := opts.BatchSize
-	if limit <= 0 {
-		limit = 100
-	}
-
-	// Scroll documents
-	scrollOpts := &types.ScrollOptions{
-		CollectionName: collectionName,
-		Filter:         filter,
-		Limit:          limit,
-		ScrollID:       opts.ScrollID,
-		OrderBy:        opts.OrderBy,
-		Fields:         opts.Fields,
-		IncludeVector:  false,
-		IncludePayload: true,
-	}
-
-	result, err := g.Vector.ScrollDocuments(ctx, scrollOpts)
-	if err != nil {
-		return nil, "", false, fmt.Errorf("failed to scroll documents: %w", err)
-	}
-
-	// Filter out nil chunks
-	var validChunks []*types.Document
-	for _, chunk := range result.Documents {
-		if chunk != nil {
-			validChunks = append(validChunks, chunk)
-		}
-	}
-
-	return validChunks, result.ScrollID, result.HasMore, nil
 }
 
 // queryNodesAndRelationshipsFromGraph queries nodes and relationships from graph database
@@ -2122,171 +706,4 @@ func (g *GraphRag) queryMetadataFromStoreOnly(docID string, segmentIDs []string)
 	}
 
 	return storeData
-}
-
-// assembleSegments assembles segments from the queried data
-func (g *GraphRag) assembleSegments(data *segmentQueryResult, graphName string, docID string) []types.Segment {
-	var segments []types.Segment
-
-	// Create segments from chunks
-	for _, chunk := range data.Chunks {
-		if chunk == nil {
-			continue
-		}
-
-		segment := types.Segment{
-			CollectionID:  graphName,
-			DocumentID:    docID,
-			ID:            chunk.ID,
-			Text:          chunk.Content,
-			Metadata:      chunk.Metadata,
-			Nodes:         []types.GraphNode{},
-			Relationships: []types.GraphRelationship{},
-			Parents:       []string{},
-			Children:      []string{},
-			Version:       1,
-			Weight:        0.0,
-			Score:         0.0,
-			Positive:      0,
-			Negative:      0,
-			Hit:           0,
-		}
-
-		// Set timestamps from metadata if available
-		if createdAt, ok := chunk.Metadata["created_at"]; ok {
-			if createdAtStr, ok := createdAt.(string); ok {
-				if t, err := time.Parse(time.RFC3339, createdAtStr); err == nil {
-					segment.CreatedAt = t
-				}
-			}
-		}
-		if updatedAt, ok := chunk.Metadata["updated_at"]; ok {
-			if updatedAtStr, ok := updatedAt.(string); ok {
-				if t, err := time.Parse(time.RFC3339, updatedAtStr); err == nil {
-					segment.UpdatedAt = t
-				}
-			}
-		}
-
-		// Add nodes that are related to this segment
-		for _, node := range data.Nodes {
-			if g.isNodeRelatedToSegment(node, chunk.ID) {
-				segment.Nodes = append(segment.Nodes, node)
-			}
-		}
-
-		// Add relationships that are related to this segment
-		for _, rel := range data.Relationships {
-			if g.isRelationshipRelatedToSegment(rel, chunk.ID) {
-				segment.Relationships = append(segment.Relationships, rel)
-			}
-		}
-
-		// Extract weight/score/vote from chunk metadata (Vector DB data) using safe extraction
-		if chunk.Metadata != nil {
-			segment.Weight = types.SafeExtractFloat64(chunk.Metadata["weight"], segment.Weight)
-			segment.Score = types.SafeExtractFloat64(chunk.Metadata["score"], segment.Score)
-			segment.Positive = types.SafeExtractInt(chunk.Metadata["positive"], segment.Positive)
-			segment.Negative = types.SafeExtractInt(chunk.Metadata["negative"], segment.Negative)
-			segment.Hit = types.SafeExtractInt(chunk.Metadata["hit"], segment.Hit)
-
-			// Extract score dimensions if available
-			if scoreDimensions, ok := chunk.Metadata["score_dimensions"]; ok {
-				if dimensionsMap, ok := scoreDimensions.(map[string]interface{}); ok {
-					segment.ScoreDimensions = make(map[string]float64)
-					for key, value := range dimensionsMap {
-						segment.ScoreDimensions[key] = types.SafeExtractFloat64(value, 0.0)
-					}
-				}
-			}
-
-			// Remove these fields from metadata since they're now in external fields
-			delete(segment.Metadata, "weight")
-			delete(segment.Metadata, "score")
-			delete(segment.Metadata, "score_dimensions")
-			delete(segment.Metadata, "positive")
-			delete(segment.Metadata, "negative")
-			delete(segment.Metadata, "hit")
-			// Note: vote is not stored in Vector DB - it's a list stored in Store only
-		}
-
-		// Also add metadata from store (for backward compatibility if Store is configured)
-		if segmentData, ok := data.StoreData[chunk.ID]; ok {
-			if segmentMap, ok := segmentData.(map[string]interface{}); ok {
-				// Store data takes precedence over Vector DB data if both exist, using safe extraction
-				segment.Weight = types.SafeExtractFloat64(segmentMap["weight"], segment.Weight)
-				segment.Score = types.SafeExtractFloat64(segmentMap["score"], segment.Score)
-				segment.Positive = types.SafeExtractInt(segmentMap["positive"], segment.Positive)
-				segment.Negative = types.SafeExtractInt(segmentMap["negative"], segment.Negative)
-				segment.Hit = types.SafeExtractInt(segmentMap["hit"], segment.Hit)
-
-				// Extract score dimensions from store if available
-				if scoreDimensions, ok := segmentMap["score_dimensions"]; ok {
-					if dimensionsMap, ok := scoreDimensions.(map[string]interface{}); ok {
-						segment.ScoreDimensions = make(map[string]float64)
-						for key, value := range dimensionsMap {
-							segment.ScoreDimensions[key] = types.SafeExtractFloat64(value, 0.0)
-						}
-					}
-				}
-
-				// Ensure these fields are not duplicated in metadata
-				delete(segment.Metadata, "weight")
-				delete(segment.Metadata, "score")
-				delete(segment.Metadata, "score_dimensions")
-				delete(segment.Metadata, "positive")
-				delete(segment.Metadata, "negative")
-				delete(segment.Metadata, "hit")
-				// Note: vote is not a single value - it's a list stored in Store only
-			}
-		}
-
-		segments = append(segments, segment)
-	}
-
-	return segments
-}
-
-// isNodeRelatedToSegment checks if a node is related to a segment
-func (g *GraphRag) isNodeRelatedToSegment(node types.GraphNode, segmentID string) bool {
-	// Check if the segment ID is in the node's source_chunks
-	if sourceChunks, ok := node.Properties["source_chunks"]; ok {
-		switch sourceChunksValue := sourceChunks.(type) {
-		case []interface{}:
-			for _, chunk := range sourceChunksValue {
-				if chunkStr, ok := chunk.(string); ok && chunkStr == segmentID {
-					return true
-				}
-			}
-		case []string:
-			for _, chunk := range sourceChunksValue {
-				if chunk == segmentID {
-					return true
-				}
-			}
-		}
-	}
-	return false
-}
-
-// isRelationshipRelatedToSegment checks if a relationship is related to a segment
-func (g *GraphRag) isRelationshipRelatedToSegment(rel types.GraphRelationship, segmentID string) bool {
-	// Check if the segment ID is in the relationship's source_chunks
-	if sourceChunks, ok := rel.Properties["source_chunks"]; ok {
-		switch sourceChunksValue := sourceChunks.(type) {
-		case []interface{}:
-			for _, chunk := range sourceChunksValue {
-				if chunkStr, ok := chunk.(string); ok && chunkStr == segmentID {
-					return true
-				}
-			}
-		case []string:
-			for _, chunk := range sourceChunksValue {
-				if chunk == segmentID {
-					return true
-				}
-			}
-		}
-	}
-	return false
 }
